refactor: extract runtime bootstrap and reconfigurator
